Fix missing header KSCrashReportWriter.h
This resolves an issue introduced in 43683817. BugsnagConfiguration.h (a
public header) gained a new include of the KSCrashReportWriter.h
(private) header. This fix includes updates to the podpsec so that
KSCrashReportWriter.h is now public, and also reverts the project
setting "CLANG_ALLOW_NON_MODULAR_INCLUDES_IN_FRAMEWORK_MODULES" which I
assume was set to allow the test target to compile.

- Add public header in Bugsnag.podspec
- Revert project setting allowing non modular includes
- Add public header in Bugsnag target