tests: Add app crash scenarios to end-to-end suite (#254)
Changes the launch incantation to pre-install the test harness app
and cause a crash prior to the scenarios being run.

The Swift scenario is unstable, dependent on both Swift version and
Xcode version. Later method name changes are likely a difference 
between Swift 4.0 and Swift 4.1, which we will need to address when 
upgrading Xcode versions.