fix: Adds fallback dictionary when serialising device info (#279)
Previously the device tab was serialised using the dictionary supplied from
the report, which could be null if the keypath did not exist. This could
result in the device tab being completely null. This fix uses an empty
dictionary as a fallback so that at least some information (freeDisk,
locale, simulator) will always be returned.

### Changed
Initialise empty dictionary and defensively add items from report to it.