fix(backtrace): Strip pointer auth mask before returning usable frames
The Arm64e architecture introduces pointer authentication codes to
detect and guard against unexpected changes to pointers in memory. For
most application functions, this is a nice bonus that nobody needs to
think about, however, for a crash reporting library, we need to strip
this extraneous value ahead of time to avoid reporting a stack frame
address which does not align to a function exactly, causing unreadable
stack traces.

PACs work by adding a signature to the higher order bits of a pointer
before it is stored. When the pointer is read, the signature is
validated prior to executing a function. If the signature has been
tampered with, then the app is forced to crash rather than execute
altered code.