fix: Check for null image name before comparing strings
If the number of loaded images changes between when _dyld_image_count()
is computed and when _dyld_get_image_name() is called, there can be less
images than the count suggests. In this case, _dyld_get_image_name()
returns NULL, and the lack of a check would crash the app (again?)
because both arguments to strcmp are required to be non-null. The
resulting exception backtrace looks somewhat like this:

```
_platform_strcmp
bsg_ksdlimageNamed (BSG_KSDynamicLinker.c:41:21)
bsg_ksdlimageUUID (BSG_KSDynamicLinker.c:56:31)
+[BSG_KSSystemInfo appUUID] (BSG_KSSystemInfo.m:158:13)
+[BSG_KSSystemInfo systemInfo] (BSG_KSSystemInfo.m:432:5)
-[BugsnagSessionTrackingPayload toJson] (BugsnagSessionTrackingPayload.m:48:5)
```