Add a high performance, async-safe key-value store for use in time sensitive code.
The idea is to minimize the chances of a KV write operation getting aborted if the app happens to crash or shut down, by making use of syscalls for the heavy lifting (syscalls can't be interrupted). The majority of the work (and thus CPU time) happens in the kernel, giving a very good chance that the write will always succeed.