Fix test flakes in enabled_error_types.feature
Occasionally the tests would fail due to a session payload being sent
twice.

In these cases the notifier was not receiving the response to its HTTP
request before the scenario crashed the app, but the server had
received it. When Bugsnag was reinitialised it would re-send the first
session since it could not know that it had been already been received
by the server.

By removing the 2 second delay (which is not always enough time to
receive the HTTP response) we can be sure the session from the first
launch of the app has not been sent, and avoid duplicates.