Fix uninstalling signal handlers
When uninstalling signal handlers disable them instead of uninstalling.
Do it this way because:

1. Signal handlers installed after ours can take our signal handling
function address and chain to it. In this case the function must always
work.
2. If multiple signal handlers for same signal are installed after ours
we can't remove ours without removing all but the last installed one.

Because we don't uninstall our signal handlers we now directly call
previously installed signal handlers.